Not sufficient to cast int to size_t where loops are decrementing. Change size_t's to long int to avoid compiler warnings while handling of decrements go negative. I tested this and can't recreate the errors Mike is getting, but that has been happening all day, so not sure if this will fix his issues.

+++ class/DataFile.h
@@ -62,7 +62,7 @@ private:
   std::string	fileName;
   std::string	projName, flightNum, flightDate;
 
-  size_t	nProbes;
+  long int	nProbes;
 
   FlightClock	startTime, endTime;
